refactored and changes as per feedback
